Add --overwrite and --cleanup options to 'saveas'
Some basic invocation of the `:saveas` command in Tridactyl could look
like the following:

- Example 1: `:saveas ~/Documents/msg.doc` Example 2: `:saveas
- ~/Documents`

In [1], `saveas` attempts to save a file called `msg.doc` inside
`$HOME/Documents` directory.

In [2], `saveas` attempts to save whatever file-name is provided by the
current URL and tries to save it in the `~/Documents` directory. In this
example, it is possible that the user tries to download another document
called `msg.doc` from the current URL (which is different than the
existing one in `~/Documents`) and unknowingly tries to save it inside
the `~/Documents` directory resulting into a conflict. This can now be
avoided with the new `--overwrite` option described below.

- Example 3: `saveas --overwrite ~/Documents/msg.doc`
- Example 4: `saveas --overwrite ~/Documents`

If user specifies `--overwrite` as shown in example [3] and [4] either
with a file or directory target, the document downloaded from the
current URL would overwrite the target file e.g. inside `~/Docuemnts`.

- Example 5: `saveas --cleanup ~/Documents/msg.doc`
- Example 6: `saveas --cleanup ~/Documents`

Similarly if `--cleanup` is specified, if the download and save/move
operation failed for any reason e.g. target file exists, the downloaded
file will be deleted from user's default downloads directory (usually
`$HOME/Downloads`). By default, the downloaded files that are failed to
be saved in the user specified location are not deleted and subsequent
`saveas` invocations on the same URL creates files like `msg.doc(n)` and
these are spuriously saved in the user specified location cluttering up
target directory and disk-space.

Hope this helps. Thank you.