Implement clone startup and waiting
Only thing left to do for a basic working version is the actual
restarting.
The original plan for the clone to run getppid twice to find out
Firefox's PID had to be scratched because the parent messenger would've
already exited by that point. This would've caused its own PID to go stale
and made it impossible to use it to find Firefox. Instead, the parent
now runs getppid once and passes the result to the clone.