Add "run_async" native command
To fix updating on Windows (#34), we need to invoke the install/update
script so that it runs whilst the messenger has already exited, so that
it can overwrite the binary. Instead of making a specialised command
just for that, I decided to just generalise it, since this seems useful
to have anyway.
This command can be used pretty much the same way as "run", but it
returns immediately and doesn't return any output or exit code. Because
of this, Tridactly cannot know whether the invoked shell command
succeeded.